functions/requirements.txt
firebase-functions~=0.1.0
flask~=3.0.0
flask-cors~=4.0.0
google-genai~=0.3.0
python-dotenv~=1.0.0
firebase-admin~=6.2.0
PyPDF2>=3.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
lxml>=4.9.0
bleach>=6.0.0
